requirements.txt
torch
torchvision
ultralytics
streamlit
pandas
Pillow
requests
numpy
